build: explicitely remove line length from docs md files
If your IDE of choice has a default line length then this may cause
auto-formatting to wrap lines in the doc markdown files even though
no max line length is set in the editor config. For the docs this
actually means something as 11ty converts the line break into a <br>
(or similar). This change makes it explicit that any line length
is ok in the doc markdown files to protect against this.